+++ internal/debug_kqueue.go
@@ -20,8 +20,6 @@ func Debug(name string, kevent *unix.Kevent_t) {
 			l = append(l, n.n)
 		}
 	}
-
-	fmt.Fprintf(os.Stderr, "%s  %-20s → %s\n",
-		time.Now().Format("15:04:05.0000"),
-		strings.Join(l, " | "), name)
+	fmt.Fprintf(os.Stderr, "FSNOTIFY_DEBUG: %s  %10d:%-60s → %q\n",
+		time.Now().Format("15:04:05.000000000"), mask, strings.Join(l, " | "), name)
 }